requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
qrcode[pil]==7.4.2
Pillow==10.4.0
